Fix asserts to reflect actual value
The stop-gap change made in an earlier revision of these tests
meant that we could only rely on a partial version of the date to
be accurate.  Since this ticket accurately reflects the real date
when the records are created, we can revert to checking the actual
string we expect to see, rather than just part of it.