Add time zone support
Date and time were being taken 'as is' from what HTCondor gives in the
log files;  metrics gives a hint at what the year is (which isn't
available in log files), but there was no way of determining time
zones.   In some edge cases (close the end of the year and in GMT),
lack of a time zone on the executing machine resulted in incorrect
times listed in the future during comparison (seconds vs datetime).
This fix adds the local time zone in all cases.   This also changes a
badly named property.